+++ include/viua/front/asm.h
@@ -30,7 +30,7 @@ struct srcline_t {
 };
 
 
-std::vector<std::string> expandSource(const std::vector<std::string>&, std::map<unsigned, unsigned>&);
+std::vector<std::string> expandSource(const std::vector<std::string>&, std::map<long unsigned, long unsigned>&);
 
 std::vector<std::vector<std::string>> decode_line_tokens(const std::vector<std::string>&);
 std::vector<std::vector<std::string>> decode_line(const std::string&);
@@ -38,7 +38,7 @@ std::vector<std::vector<std::string>> decode_line(const std::string&);
 int gatherFunctions(invocables_t*, const std::vector<std::string>&, const std::vector<std::string>&);
 int gatherBlocks(invocables_t*, const std::vector<std::string>&, const std::vector<std::string>&);
 
-int generate(const std::vector<std::string>&, const std::map<unsigned, unsigned>&, std::vector<std::string>&, invocables_t&, invocables_t&, std::string&, std::string&, const std::vector<std::string>&, const compilationflags_t&);
+int generate(const std::vector<std::string>&, const std::map<long unsigned, long unsigned>&, std::vector<std::string>&, invocables_t&, invocables_t&, std::string&, std::string&, const std::vector<std::string>&, const compilationflags_t&);
 
 
 #endif
